+++ pex/resolver.py
@@ -1545,6 +1545,7 @@ def resolve(
     use_pip_config=False,  # type: bool
     extra_pip_requirements=(),  # type: Tuple[Requirement, ...]
     keyring_provider=None,  # type: Optional[str]
+    uploaded_prior_to=None,  # type: Optional[str]
     result_type=InstallableType.INSTALLED_WHEEL_CHROOT,  # type: InstallableType.Value
     dependency_configuration=DependencyConfiguration(),  # type: DependencyConfiguration
 ):
@@ -1640,6 +1641,7 @@ def resolve(
         use_pip_config=use_pip_config,
         extra_pip_requirements=extra_pip_requirements,
         keyring_provider=keyring_provider,
+        uploaded_prior_to=uploaded_prior_to,
     )
 
     requests = tuple(
@@ -1936,6 +1938,7 @@ def download_requests(
     use_pip_config=False,  # type: bool
     extra_pip_requirements=(),  # type: Tuple[Requirement, ...]
     keyring_provider=None,  # type: Optional[str]
+    uploaded_prior_to=None,  # type: Optional[str]
     dependency_configuration=DependencyConfiguration(),  # type: DependencyConfiguration
 ):
     # type: (...) -> Downloaded
@@ -1948,6 +1951,7 @@ def download_requests(
         use_pip_config=use_pip_config,
         extra_pip_requirements=extra_pip_requirements,
         keyring_provider=keyring_provider,
+        uploaded_prior_to=uploaded_prior_to,
     )
 
     build_requests, download_results = _download_internal(
@@ -2023,6 +2027,7 @@ def reports(
     use_pip_config=False,  # type: bool
     extra_pip_requirements=(),  # type: Tuple[Requirement, ...]
     keyring_provider=None,  # type: Optional[str]
+    uploaded_prior_to=None,  # type: Optional[str]
     dependency_configuration=DependencyConfiguration(),  # type: DependencyConfiguration
 ):
     # type: (...) -> Reports
@@ -2035,6 +2040,7 @@ def reports(
         use_pip_config=use_pip_config,
         extra_pip_requirements=extra_pip_requirements,
         keyring_provider=keyring_provider,
+        uploaded_prior_to=uploaded_prior_to,
     )
 
     pip_session = _PipSession(